fix: Eliminate all FutureWarnings and update STATUS.md
- Fix pct_change() FutureWarning in core/regime_filter.py (add fill_method=None)
- Fix pct_change() FutureWarning in ml_features/regime_ml.py (add fill_method=None)
- Fix dtype FutureWarning in tests/unit/test_ml_features.py (explicit int cast)
- Update docs/STATUS.md with 869 tests passing, 0 warnings status

All tests now pass with zero warnings:
  869 passed, 4 skipped (tkinter-dependent), 0 warnings

+++ tests/unit/test_ml_features.py
@@ -220,8 +220,8 @@ class TestAnomalyDetection:
         detector = AnomalyDetector()
 
         df = create_ohlcv_data(num_days=100)
-        # Add a volume spike
-        df.loc[80, 'volume'] = df['volume'].mean() * 10
+        # Add a volume spike (explicit cast to avoid dtype warning)
+        df.loc[80, 'volume'] = int(df['volume'].mean() * 10)
 
         result = detector.detect_all(df)
